logic abstracted to javscript file
